project/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.11
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-timezone-field==7.0
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.7
kombu==5.4.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
prompt_toolkit==3.0.47
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
